remove global flag when testing regex
Global flag: When the regular expression has the global flag /g, the .test() method will search for all matches, not just the first one. The global flag tells the regular expression engine to start the search at the last match instead of the beginning of the string. This means that if you test the same value multiple times with a global regular expression, you may get different results depending on where the last match was found.